13/11
Eerste aanzet to admin ==> admin in db enkel het navigeren naar de admin homepage is een gedoe en werkt nog totaal niet

+++ database/migrations/2024_11_05_111742_add_username_to_users_table.php
@@ -1,21 +1,25 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+class AddUsernameToUsersTable extends Migration
+{
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('username')->unique()->after('email'); // Voeg de 'username' kolom toe
+            if (!Schema::hasColumn('users', 'username')) {
+                $table->string('username')->nullable()->after('email');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('username'); // Verwijder de 'username' kolom bij rollback
+            if (Schema::hasColumn('users', 'username')) {
+                $table->dropColumn('username');
+            }
         });
     }
-};
+}
