+++ database/migrations/2024_11_05_124705_add_description_to_users_table.php
@@ -1,23 +1,25 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class AddDescriptionToUsersTable extends Migration
 {
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->text('description')->nullable()->after('profile_picture'); // Voeg deze regel toe
+            if (!Schema::hasColumn('users', 'description')) {
+                $table->text('description')->nullable()->after('profile_picture');
+            }
         });
     }
-    
+
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('description'); // Voeg deze regel toe
+            if (Schema::hasColumn('users', 'description')) {
+                $table->dropColumn('description');
+            }
         });
     }
-    
-};
+}
